+++ consensus/state_sync/helpers.go
@@ -12,13 +12,7 @@ func (m *stateSync) SendStateSyncMessage(stateSyncMsg *typesCons.StateSyncMessag
 		return err
 	}
 
-	fields := map[string]any{
-		"height":     height,
-		"peerId":     peerId,
-		"proto_type": getMessageType(stateSyncMsg),
-	}
-
-	m.logger.Info().Fields(fields).Msg("Sending StateSync Message")
+	m.logger.Info().Fields(m.logHelper(string(peerId))).Msg("Sending StateSync Message")
 	return m.sendToPeer(anyMsg, peerId)
 }
 
@@ -44,4 +38,15 @@ func getMessageType(msg *typesCons.StateSyncMessage) string {
 	default:
 		return "Unknown"
 	}
+}
+
+func (m *stateSync) logHelper(receiver string) map[string]any {
+	consensusMod := m.GetBus().GetConsensusModule()
+
+	return map[string]any{
+		"height":   consensusMod.CurrentHeight(),
+		"sender":   consensusMod.GetNodeAddress(),
+		"receiver": receiver,
+	}
+
 }