+++ p2p/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.34] - 2023-03-09
+## [0.0.0.34] - 2023-03-10
 
 - Add log for `StateMachineTransitionEvent`